requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
alembic==1.13.2
redis==5.0.7
prometheus-client==0.20.0
psycopg2-binary==2.9.9
